feat: component cleanup and use string hash colors
