+++ package.json
@@ -18,6 +18,7 @@
     "chart.js": "^3.7.0",
     "chartjs-plugin-annotation": "^1.3.1",
     "core-js": "^3.21.0",
+    "md5": "^2.3.0",
     "nuxt": "^2.15.8",
     "ol": "^6.12.0",
     "vue": "^2.6.14",
